Loto Settings Detail page was failing to delete a non-existent LotoSettings JPA Entity.  We now check to see that this entity has a valid ID before attempting to delete it.

+++ fid-web/src/main/java/com/n4systems/fieldid/wicket/pages/setup/loto/LotoDetailsSetupPage.java
@@ -62,8 +62,12 @@ public class LotoDetailsSetupPage extends FieldIDTemplatePage {
 
                 if(customLotoDetails.getApplicationProcess() == null &&
                         customLotoDetails.getRemovalProcess() == null &&
-                        customLotoDetails.getTestingAndVerification() == null ) {
-                    procedureDefinitionService.deleteLotoSettings(customLotoDetails);
+                        customLotoDetails.getTestingAndVerification() == null) {
+
+                    //If the ID is null, it doesn't exist... so we can't delete it.
+                    if(customLotoDetails.getId() != null) {
+                        procedureDefinitionService.deleteLotoSettings(customLotoDetails);
+                    }
                 } else {
                     procedureDefinitionService.saveOrUpdateLotoSettings(customLotoDetails);
                 }
